requirements.txt
pygame
opencv-python-headless
gtts
playsound==1.2.2
pyserial
torch
ultralytics
